New functionality: dry run of each command
