+++ Permissions/permissions/commands.py
@@ -6,7 +6,7 @@ import os
 import subprocess
 
 from permissions import exceptions
-from permissions.utils import log_debug
+from permissions.utils import log
 
 logger = logging.getLogger(__name__)
 
@@ -19,7 +19,7 @@ class BaseCommand(object):
     should_return_value = False
     should_stream_output = False
 
-    @log_debug(logger)
+    @log(logger)
     def prepare(self, *args, **kwargs):
         for param in self.command_params:
             if param not in kwargs:
@@ -28,8 +28,11 @@ class BaseCommand(object):
                 )
         self.command = self.command.format(**kwargs)
 
-    @log_debug(logger)
-    def execute(self, *args, **kwargs):
+    @log(logger)
+    def execute(self, dry_run=False, **kwargs):
+        if dry_run:
+            return
+
         output = []
 
         proc = subprocess.Popen(self.command.split(), stdout=subprocess.PIPE, bufsize=1)
@@ -45,14 +48,14 @@ class BaseCommand(object):
         if self.should_return_value:
             return ''.join(output)
 
-    @log_debug(logger)
+    @log(logger)
     def collect_artifacts(self, *args, **kwargs):
         if not self.should_gather_artifacts:
             raise exceptions.CommandNotArtifactableException(
                 "Command is not set to be artifactable."
             )
 
-    @log_debug(logger)
+    @log(logger)
     def clean(self, *args, **kwargs):
         raise NotImplementedError()
 
@@ -113,41 +116,3 @@ class CountedPingCommand(BaseCommand):
 
 class DummyCommand(BaseCommand):
     command = 'echo Dummy'
-
-
-class CommandManager(object):
-    """
-    A simple command manager that makes it easy to execute command
-    without getting into its definition, just based on a knowledge
-    that each command needs parameters. So basically, you need to
-    know what parameters the command is expecting and pass it into
-    `set_parameters` method. But even if you don't, meaningful
-    exceptions will be raised.
-
-    Example of usage:
-
-    cm = CommandManager()
-    cm.set_command(CountedPingCommand())
-    cm.set_parameters(host='www.google.com', count=3)
-    cm.run()
-    """
-
-    def __init__(self):
-        self._command = None
-        self._parameters = dict()
-
-    def set_command(self, command):
-        if not isinstance(command, BaseCommand):
-            raise exceptions.NotACommandException("Command should be an instance of BaseCommand.")
-
-        self._command = command
-
-    def set_parameters(self, **params):
-        self._parameters = params
-
-    def run(self):
-        if not self._command:
-            raise exceptions.NoCommandSetException("You should set_command first.")
-
-        self._command.prepare(**self._parameters)
-        self._command.execute()
